Move client into Client folder, add PowerShell and Bash script to publish and prerender

+++ PublishAndPrerender.ps1
@@ -1,5 +1,14 @@
-Remove-Item -Path .\Prerender\output -Recurse
-dotnet publish -c Release -o Prerender/output
+If(Test-Path .\Prerender\output)
+{
+    Remove-Item -Path .\Prerender\output -Recurse
+}
+
+dotnet publish .\src\BlazorWasmPrerender.csproj -c Release -o Prerender/output
 Push-Location .\Prerender
 npx react-snap
+Get-ChildItem ".\output\wwwroot\*.html" -Recurse | ForEach-Object { 
+    (Get-Content -Path $_.FullName -Raw) `
+        -replace '<base href="/"','<base href="/GitHubPagesDemo/"' | `
+        Set-Content -Path $_.FullName
+}
 Pop-Location
